+++ admin/class-tmsm-aquatonic-course-booking-admin.php
@@ -72,18 +72,6 @@ class Tmsm_Aquatonic_Course_Booking_Admin {
 	 */
 	public function enqueue_styles() {
 
-		/**
-		 * This function is provided for demonstration purposes only.
-		 *
-		 * An instance of this class should be passed to the run() function
-		 * defined in Tmsm_Aquatonic_Course_Booking_Loader as all of the hooks are defined
-		 * in that particular class.
-		 *
-		 * The Tmsm_Aquatonic_Course_Booking_Loader will then create the relationship
-		 * between the defined hooks and the functions defined in this
-		 * class.
-		 */
-
 		wp_enqueue_style( $this->plugin_name, plugin_dir_url( __FILE__ ) . 'css/tmsm-aquatonic-course-booking-admin.css', array(), $this->version, 'all' );
 
 	}
@@ -95,19 +83,8 @@ class Tmsm_Aquatonic_Course_Booking_Admin {
 	 */
 	public function enqueue_scripts() {
 
-		/**
-		 * This function is provided for demonstration purposes only.
-		 *
-		 * An instance of this class should be passed to the run() function
-		 * defined in Tmsm_Aquatonic_Course_Booking_Loader as all of the hooks are defined
-		 * in that particular class.
-		 *
-		 * The Tmsm_Aquatonic_Course_Booking_Loader will then create the relationship
-		 * between the defined hooks and the functions defined in this
-		 * class.
-		 */
-
-		wp_enqueue_script( $this->plugin_name, plugin_dir_url( __FILE__ ) . 'js/tmsm-aquatonic-course-booking-admin.js', array( 'jquery' ), $this->version, true );
+		wp_enqueue_script( 'jquery.countdown', plugin_dir_url( __FILE__ ) . 'js/jquery.countdown.min.js', array( 'jquery' ), $this->version, true );
+		wp_enqueue_script( $this->plugin_name, plugin_dir_url( __FILE__ ) . 'js/tmsm-aquatonic-course-booking-admin.js', array( 'jquery', 'jquery.countdown' ), $this->version, true );
 
 	}